adjust profile page data
